refactor: Checker -> AttendanceService
